Resolve migration issues
For whatever reason, the initial create_student migration was modified to also add several other columns than it was initially. One of these was paper_history, leading to the later migration for creating paper_history causing a failure due to the column already existing. I can't see if this works on heroku without commiting this, so here we go.