Attempt to add new defaults for student entries
HOPEFULLY this will allow running on heroku. We'll see, I suppose. The option if we were actually running a production site would be to find a way to force new defaults on entries. Seems like Rails doesn't have that functionality built-in, though, so we're gonna try just this for now.